+++ src/com/ansorgit/plugins/bash/lang/parser/arithmetic/AbstractRepeatedExpr.java
@@ -2,13 +2,13 @@
  * Copyright 2011 Joachim Ansorg, mail@ansorg-it.com
  * File: AbstractRepeatedExpr.java, Class: AbstractRepeatedExpr
  * Last modified: 2011-04-30 16:33
- *
+ * <p/>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p/>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p/>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -35,14 +35,24 @@ import com.intellij.psi.tree.TokenSet;
 class AbstractRepeatedExpr implements ArithmeticParsingFunction {
     private final ArithmeticParsingFunction expressionParser;
     private final TokenSet operators;
+    private boolean prefixOperator;
     private final IElementType partMarker;
     private int maxRepeats;
     private final String debugInfo;
     private final boolean checkWhitespace;
 
-    AbstractRepeatedExpr(ArithmeticParsingFunction expressionParser, TokenSet operators, IElementType partMarker, int maxRepeats, String debugInfo) {
+    /**
+     * @param expressionParser
+     * @param operators
+     * @param prefixOperator   true if the operators are expected to be before the subexpression, otherweise between subexpressions
+     * @param partMarker
+     * @param maxRepeats
+     * @param debugInfo
+     */
+    AbstractRepeatedExpr(ArithmeticParsingFunction expressionParser, TokenSet operators, boolean prefixOperator, IElementType partMarker, int maxRepeats, String debugInfo) {
         this.expressionParser = expressionParser;
         this.operators = operators;
+        this.prefixOperator = prefixOperator;
         this.partMarker = partMarker;
         this.maxRepeats = maxRepeats;
         this.debugInfo = debugInfo;
@@ -51,10 +61,16 @@ class AbstractRepeatedExpr implements ArithmeticParsingFunction {
     }
 
     public boolean isValid(BashPsiBuilder builder) {
-        if (expressionParser.isValid(builder)) {
+        if (prefixOperator && operators.contains(builder.getTokenType(checkWhitespace))) {
+            return true;
+        } else if (expressionParser.isValid(builder)) {
             return true;
         }
 
+        return isValidParentesisExpression(builder);
+    }
+
+    private boolean isValidParentesisExpression(BashPsiBuilder builder) {
         PsiBuilder.Marker marker = builder.mark();
 
         ArithmeticParsingFunction parenthesisParser = ArithmeticFactory.parenthesisParser();
@@ -63,11 +79,42 @@ class AbstractRepeatedExpr implements ArithmeticParsingFunction {
                 && operators.contains(builder.getTokenType(checkWhitespace));
 
         marker.rollbackTo();
-
         return ok;
     }
 
     public boolean parse(BashPsiBuilder builder) {
+        return prefixOperator
+                ? parseWithPrefixOperator(builder)
+                : parseWithNonPrefixOperator(builder);
+    }
+
+    private boolean parseWithPrefixOperator(BashPsiBuilder builder) {
+        PsiBuilder.Marker marker = builder.mark();
+
+        int count = 0;
+        boolean ok;
+
+        do {
+            ok = ParserUtil.conditionalRead(builder, operators);
+            count++;
+        } while (ok && (maxRepeats <= 0 || count < maxRepeats));
+
+        if (expressionParser.isValid(builder)) {
+            ok = expressionParser.parse(builder);
+        } else {
+            ok = ArithmeticFactory.parenthesisParser().parse(builder);
+        }
+
+        if (ok && count > 1 && partMarker != null) {
+            marker.done(partMarker);
+        } else {
+            marker.drop();
+        }
+
+        return ok;
+    }
+
+    private boolean parseWithNonPrefixOperator(BashPsiBuilder builder) {
         PsiBuilder.Marker marker = builder.mark();
 
         int count = 0;
@@ -83,7 +130,7 @@ class AbstractRepeatedExpr implements ArithmeticParsingFunction {
             count++;
         } while (ok && (maxRepeats <= 0 || count < maxRepeats) && ParserUtil.conditionalRead(builder, operators));
 
-        if (count > 1 && partMarker != null) {
+        if (ok && count > 1 && partMarker != null) {
             marker.done(partMarker);
         } else {
             marker.drop();